Add files via upload
This web scraper is built with Go and has an optional feature to search for a case-sensitive, single-word, string value. This search confirms on which lines in the .html file the given search word appears in. In order to enable this feature, simply uncomment the last line of the main.go file. The search algorithm only works with case-sensitive, single-word searches because strings in Go are implicitly immutable.